Improve the `parseRequestBody` function and all nested functions.

Original benchmarks:
```shell
% go test -benchmem -bench=. -run=^Benchmark
goos: darwin
goarch: amd64
pkg: github.com/go-resty/resty/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_parseRequestBody_string-16       2199196    550.3 ns/op     128 B/op      3 allocs/op
Benchmark_parseRequestBody_byte-16         2264421    532.9 ns/op     128 B/op      3 allocs/op
Benchmark_parseRequestBody_reader-16       8307141    141.8 ns/op      16 B/op      1 allocs/op
Benchmark_parseRequestBody_struct-16        931632     1317 ns/op     156 B/op      5 allocs/op
Benchmark_parseRequestBody_struct_xml-16    409074     2921 ns/op    4765 B/op     13 allocs/op
Benchmark_parseRequestBody_map-16           566750     2158 ns/op     570 B/op     15 allocs/op
Benchmark_parseRequestBody_slice-16         957828     1279 ns/op     146 B/op      4 allocs/op
Benchmark_parseRequestBody_FormData-16      954213     1266 ns/op     304 B/op     14 allocs/op
Benchmark_parseRequestBody_MultiPart-16      94276    12538 ns/op    8746 B/op    131 allocs/op
```

After the improvements:
```shell
Benchmark_parseRequestBody_string-16                              7635237    155.7 ns/op      16 B/op      1 allocs/op
Benchmark_parseRequestBody_byte-16                                8453085    140.4 ns/op      16 B/op      1 allocs/op
Benchmark_parseRequestBody_reader_with_SetContentLength-16       17494558    67.62 ns/op      16 B/op      1 allocs/op
Benchmark_parseRequestBody_reader_without_SetContentLength-16    26582679    45.11 ns/op       0 B/op      0 allocs/op
Benchmark_parseRequestBody_struct-16                              1264702    943.9 ns/op      40 B/op      2 allocs/op
Benchmark_parseRequestBody_struct_xml-16                           479361     2496 ns/op    4645 B/op     10 allocs/op
Benchmark_parseRequestBody_map-16                                  685204     1770 ns/op     454 B/op     12 allocs/op
Benchmark_parseRequestBody_slice-16                               1318749    910.2 ns/op      32 B/op      2 allocs/op
Benchmark_parseRequestBody_FormData-16                            1000000     1116 ns/op     272 B/op      9 allocs/op
Benchmark_parseRequestBody_MultiPart-16                             92143    12532 ns/op    8738 B/op    130 allocs/op
```