@moorereason 2023-09-29
Tip: use [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat).