@SVilgelm 2023-09-29
```shell
% benchstat old.txt new.txt 
goos: darwin
goarch: amd64
pkg: github.com/go-resty/resty/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
                                                     │    old.txt    │                new.txt                │
                                                     │    sec/op     │    sec/op     vs base                 │
_parseRequestHeader-16                                  159.6n ± ∞ ¹   152.4n ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_string-16                             550.2n ± ∞ ¹   156.8n ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_byte-16                               531.5n ± ∞ ¹   142.1n ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_reader_with_SetContentLength-16      141.40n ± ∞ ¹   69.34n ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_reader_without_SetContentLength-16    84.30n ± ∞ ¹   44.86n ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_struct-16                            1307.0n ± ∞ ¹   953.0n ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_struct_xml-16                         2.886µ ± ∞ ¹   2.471µ ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_map-16                                2.152µ ± ∞ ¹   1.775µ ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_slice-16                             1264.0n ± ∞ ¹   913.0n ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_FormData-16                           1.243µ ± ∞ ¹   1.247µ ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_MultiPart-16                          12.32µ ± ∞ ¹   12.54µ ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                                 801.3n         511.7n        -36.13%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                                     │    old.txt    │                new.txt                 │
                                                     │     B/op      │     B/op       vs base                 │
_parseRequestHeader-16                                   0.000 ± ∞ ¹     0.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_string-16                             128.00 ± ∞ ¹     16.00 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_byte-16                               128.00 ± ∞ ¹     16.00 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_reader_with_SetContentLength-16        16.00 ± ∞ ¹     16.00 ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_reader_without_SetContentLength-16     0.000 ± ∞ ¹     0.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_struct-16                             155.00 ± ∞ ¹     40.00 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_struct_xml-16                        4.650Ki ± ∞ ¹   4.533Ki ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_map-16                                 569.0 ± ∞ ¹     454.0 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_slice-16                              146.00 ± ∞ ¹     32.00 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_FormData-16                            304.0 ± ∞ ¹     272.0 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_MultiPart-16                         8.541Ki ± ∞ ¹   8.533Ki ± ∞ ¹        ~ (p=1.000 n=1) ³
geomean                                                            ⁴                  -48.94%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ need >= 4 samples to detect a difference at alpha level 0.05
⁴ summaries must be >0 to compute geomean

                                                     │   old.txt    │               new.txt                │
                                                     │  allocs/op   │  allocs/op   vs base                 │
_parseRequestHeader-16                                  0.000 ± ∞ ¹   0.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_string-16                             3.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_byte-16                               3.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_reader_with_SetContentLength-16       1.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_reader_without_SetContentLength-16    0.000 ± ∞ ¹   0.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestBody_struct-16                             5.000 ± ∞ ¹   2.000 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_struct_xml-16                         13.00 ± ∞ ¹   10.00 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_map-16                                15.00 ± ∞ ¹   12.00 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_slice-16                              4.000 ± ∞ ¹   2.000 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_FormData-16                          14.000 ± ∞ ¹   9.000 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestBody_MultiPart-16                          131.0 ± ∞ ¹   130.0 ± ∞ ¹        ~ (p=1.000 n=1) ³
geomean                                                           ⁴                -35.02%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ need >= 4 samples to detect a difference at alpha level 0.05
⁴ summaries must be >0 to compute geomean
```