@moorereason 2023-09-29
Also, use something like `-count=10` to gather more samples. That should get rid of the benchstat footnotes.