@SVilgelm 2023-10-02
New benchmarks, also no significant changes:
```
goos: darwin
goarch: amd64
pkg: github.com/go-resty/resty/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
                                                     │    old.txt    │               new.txt               │
                                                     │    sec/op     │   sec/op     vs base                │
_parseRequestHeader-16                                  171.2n ±  3%   173.7n ± 6%   +1.46% (p=0.009 n=10)
_parseRequestBody_string-16                             542.7n ±  2%   183.3n ± 1%  -66.22% (p=0.000 n=10)
_parseRequestBody_byte-16                               536.8n ±  1%   176.3n ± 1%  -67.15% (p=0.000 n=10)
_parseRequestBody_reader_with_SetContentLength-16      161.15n ±  0%   87.97n ± 1%  -45.41% (p=0.000 n=10)
_parseRequestBody_reader_without_SetContentLength-16    89.63n ±  1%   50.53n ± 0%  -43.62% (p=0.000 n=10)
_parseRequestBody_struct-16                            1250.0n ± 10%   863.2n ± 0%  -30.94% (p=0.000 n=10)
_parseRequestBody_struct_xml-16                         2.854µ ±  3%   2.534µ ± 3%  -11.23% (p=0.000 n=10)
_parseRequestBody_map-16                                1.842µ ±  2%   1.451µ ± 0%  -21.21% (p=0.000 n=10)
_parseRequestBody_slice-16                             1166.5n ±  1%   834.8n ± 0%  -28.44% (p=0.000 n=10)
_parseRequestBody_FormData-16                           1.288µ ±  1%   1.003µ ± 1%  -22.17% (p=0.000 n=10)
_parseRequestBody_MultiPart-16                          14.65µ ±  1%   14.62µ ± 1%        ~ (p=0.183 n=10)
geomean                                                 814.1n         532.1n       -34.64%

                                                     │    old.txt     │                new.txt                 │
                                                     │      B/op      │     B/op      vs base                  │
_parseRequestHeader-16                                   0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
_parseRequestBody_string-16                             128.00 ± 0%       16.00 ± 0%  -87.50% (p=0.000 n=10)
_parseRequestBody_byte-16                               128.00 ± 0%       16.00 ± 0%  -87.50% (p=0.000 n=10)
_parseRequestBody_reader_with_SetContentLength-16        16.00 ± 0%       16.00 ± 0%        ~ (p=1.000 n=10) ¹
_parseRequestBody_reader_without_SetContentLength-16     0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
_parseRequestBody_struct-16                             155.00 ± 0%       40.00 ± 0%  -74.19% (p=0.000 n=10)
_parseRequestBody_struct_xml-16                        4.649Ki ± 0%     4.534Ki ± 0%   -2.48% (p=0.000 n=10)
_parseRequestBody_map-16                                 487.0 ± 0%       372.0 ± 0%  -23.61% (p=0.000 n=10)
_parseRequestBody_slice-16                              145.00 ± 0%       32.00 ± 0%  -77.93% (p=0.000 n=10)
_parseRequestBody_FormData-16                            304.0 ± 0%       272.0 ± 0%  -10.53% (p=0.000 n=10)
_parseRequestBody_MultiPart-16                         9.919Ki ± 0%     9.911Ki ± 0%   -0.08% (p=0.000 n=10)
geomean                                                             ²                 -49.11%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                     │    old.txt    │               new.txt                │
                                                     │   allocs/op   │ allocs/op   vs base                  │
_parseRequestHeader-16                                  0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
_parseRequestBody_string-16                             3.000 ± 0%     1.000 ± 0%  -66.67% (p=0.000 n=10)
_parseRequestBody_byte-16                               3.000 ± 0%     1.000 ± 0%  -66.67% (p=0.000 n=10)
_parseRequestBody_reader_with_SetContentLength-16       1.000 ± 0%     1.000 ± 0%        ~ (p=1.000 n=10) ¹
_parseRequestBody_reader_without_SetContentLength-16    0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
_parseRequestBody_struct-16                             5.000 ± 0%     2.000 ± 0%  -60.00% (p=0.000 n=10)
_parseRequestBody_struct_xml-16                         13.00 ± 0%     10.00 ± 0%  -23.08% (p=0.000 n=10)
_parseRequestBody_map-16                                13.00 ± 0%     10.00 ± 0%  -23.08% (p=0.000 n=10)
_parseRequestBody_slice-16                              4.000 ± 0%     2.000 ± 0%  -50.00% (p=0.000 n=10)
_parseRequestBody_FormData-16                          14.000 ± 0%     9.000 ± 0%  -35.71% (p=0.000 n=10)
_parseRequestBody_MultiPart-16                          133.0 ± 0%     132.0 ± 0%   -0.75% (p=0.000 n=10)
geomean                                                            ²               -35.25%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```